Revert "Authorize.Net: Switch to the XML API"
This reverts commit a6cfae5e77ed249de84dbe46a4a91e3a0903a11d.
