+++ CHANGELOG
@@ -1,7 +1,5 @@
 = ActiveMerchant CHANGELOG
 
-* Authorize.Net: Switch to the XML API [ksondere, ntalbott]
-
 == Version 1.44.0 (Aug 21, 2014)
 
 * Moneris: Add :avs_enabled option [bslobodin]
