+++ test/fixtures.yml
@@ -20,10 +20,9 @@ app55:
   api_key: "QDtACYMCFqtuKOQ22QtCkGR1TzD7XM8i"
   api_secret: "yT1FRpuusBIQoEBIfIQ8bPStkl7yzdTz"
 
-# Working credentials, no need to replace
 authorize_net:
-  login: 4Pf3E54zCsW
-  password: 337UTt4y6DB64ft8
+  login: X
+  password: Y
 
 # Working credentials, no need to replace
 balanced:
