+++ test/unit/gateways/secure_pay_test.rb
@@ -63,8 +63,21 @@ class SecurePayTest < Test::Unit::TestCase
     assert_nil response.cvv_result['code']
   end
 
+
+  def test_undefine_unsupported_methods
+    assert @gateway.respond_to?(:purchase)
+
+    [ :authorize, :capture, :void, :credit ].each do |m|
+      assert !@gateway.respond_to?(m)
+    end
+  end
+
   def test_supported_countries
     assert_equal %w(US CA GB AU), SecurePayGateway.supported_countries
+  end
+
+  def test_supported_card_types_are_inherited
+    assert_equal AuthorizeNetGateway.supported_cardtypes, SecurePayGateway.supported_cardtypes
   end
 
   private
